Android compiler tests moved to test package
