+++ generators/src/org/jetbrains/kotlin/generators/tests/GenerateTests.kt
@@ -140,15 +140,15 @@ import org.jetbrains.kotlin.completion.handlers.AbstractBasicCompletionHandlerTe
 import org.jetbrains.kotlin.idea.decompiler.stubBuilder.AbstractClsStubBuilderTest
 import org.jetbrains.kotlin.codegen.AbstractLineNumberTest
 import org.jetbrains.kotlin.completion.handlers.AbstractKeywordCompletionHandlerTest
-import org.jetbrains.jet.lang.resolve.android.AbstractAndroidXml2KConversionTest
 import org.jetbrains.jet.android.AbstractCrossParserTest
-import org.jetbrains.jet.lang.resolve.android.AbstractAndroidBoxTest
 import org.jetbrains.jet.android.AbstractAndroidCompletionTest
 import org.jetbrains.jet.android.AbstractAndroidGotoTest
 import org.jetbrains.jet.jps.build.android.AbstractAndroidJpsTestCase
 import org.jetbrains.jet.android.AbstractAndroidRenameTest
 import org.jetbrains.jet.android.AbstractAndroidFindUsagesTest
-import org.jetbrains.jet.lang.resolve.android.AbstractAndroidBytecodeShapeTest
+import org.jetbrains.jet.lang.resolve.android.test.AbstractAndroidBytecodeShapeTest
+import org.jetbrains.jet.lang.resolve.android.test.AbstractAndroidXml2KConversionTest
+import org.jetbrains.jet.lang.resolve.android.test.AbstractAndroidBoxTest
 
 fun main(args: Array<String>) {
     System.setProperty("java.awt.headless", "true")
