+++ plugins/android-compiler-plugin/tests/org/jetbrains/jet/lang/resolve/android/test/AndroidBytecodeShapeTestGenerated.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.jetbrains.jet.lang.resolve.android;
+package org.jetbrains.jet.lang.resolve.android.test;
 
 import com.intellij.testFramework.TestDataPath;
 import org.jetbrains.jet.JUnit3RunnerWithInners;
